Remove d.ts definitions to fix ts-node and mocha compilations
