+++ package.json
@@ -25,7 +25,7 @@
     "typescript": "^3.4.5"
   },
   "scripts": {
-    "test": "mocha -r ts-node/register ./test.ts",
+    "test": "mocha --compilers ts:ts-node/register ./test.ts",
     "compile": "tsc",
     "prepublish": "npm run compile"
   },
