+++ src/lib/PromiseStates.ts
@@ -1,0 +1,5 @@
+export enum PromiseStates {
+  Pending = 0,
+  Fulfilled = 1,
+  Rejected = 2
+}
