--- src/lib/states.ts
@@ -1,5 +1,0 @@
-export enum PromiseStates {
-  Pending,
-  Fulfilled,
-  Rejected
-}
